fix: Done 버튼 클릭 후 입력칸 비우기
