--- note-project/src/hooks/useChange.ts
@@ -1,15 +1,0 @@
-import { useState } from "react";
-
-const useChange = () => {
-  const [title, setTitle] = useState("");
-  const [content, setContent] = useState("");
-
-  const handleTitleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setTitle(e.target.value);
-  };
-  const handleContentChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setContent(e.target.value);
-  };
-  return { title, content, handleTitleChange, handleContentChange };
-};
-export default useChange;
