+++ note-project/src/hooks/useCreate.ts
@@ -1,0 +1,36 @@
+import { useState } from "react";
+import { saveNote } from "../utils/storageNote";
+
+const useCreate = () => {
+  const [title, setTitle] = useState("");
+  const [content, setContent] = useState("");
+
+  const handleTitleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    setTitle(e.target.value);
+  };
+  const handleContentChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    setContent(e.target.value);
+  };
+  const handleDoneClick = () => {
+    if (title.trim().length === 0 || content.trim().length === 0) {
+      alert("빈칸을 정확히 입력해주세요.");
+      return;
+    }
+    if (title && content) {
+      const newNote = { title, content };
+      saveNote(newNote);
+      setTitle("");
+      setContent("");
+    }
+  };
+  return {
+    title,
+    content,
+    setTitle,
+    setContent,
+    handleTitleChange,
+    handleContentChange,
+    handleDoneClick,
+  };
+};
+export default useCreate;
