+++ note-project/src/pages/Create/Create.tsx
@@ -2,24 +2,18 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faChevronLeft } from "@fortawesome/free-solid-svg-icons";
 import styles from "./Create.module.css";
 import Button from "../../components/commons/Button/Button";
-import useChange from "../../hooks/useChange";
-import { saveNote } from "../../utils/storageNote";
+import useCreate from "../../hooks/useCreate";
 
 const Create = () => {
-  const { title, content, handleTitleChange, handleContentChange } =
-    useChange();
+  const {
+    title,
+    content,
+    handleTitleChange,
+    handleContentChange,
+    handleDoneClick,
+  } = useCreate();
   const handleRemoveNote = () => {
     console.log("remove note");
-  };
-  const handleDoneClick = () => {
-    if (title.trim().length === 0 || content.trim().length === 0) {
-      alert("빈칸을 정확히 입력해주세요.");
-      return;
-    }
-    if (title && content) {
-      const newNote = { title, content };
-      saveNote(newNote);
-    }
   };
   return (
     <section className={styles.container}>
@@ -31,11 +25,13 @@ const Create = () => {
       <main className={styles.main}>
         <input
           type="text"
+          value={title}
           className={styles.title}
           placeholder="제목을 입력하세요."
           onChange={handleTitleChange}
         />
         <textarea
+          value={content}
           className={styles.content}
           placeholder="내용을 입력하세요."
           onChange={handleContentChange}
